Reorganize header files (closes #14)
group together the fast5 headers, separating (most of) them from
the example headers.
