+++ src/fast5/cwmap.fast5_fq_bp_1.inl
